+++ tasking/utils.py
@@ -140,22 +140,6 @@ def generate_task_occurrences(task):
     return task.taskoccurrence_set.all()
 
 
-def validate_rrule(rule_string):
-    """
-    Validates an rrule string; returns True or False
-    """
-    try:
-        rrulestr(rule_string)
-    except ValueError:
-        # this string is not a valid rrule
-        return False
-    except TypeError:
-        # this is not even a string
-        return False
-    else:
-        return True
-
-
 def get_rrule_start(rrule_obj):
     """
     Returns the timezone-aware start datetime from rrule
